requirements.txt
Django==4.0.5
django-bootstrap5==21.3
djangorestframework~=3.13.1
